pyproject.toml
[project]
name = "ai-research-template"
version = "0.1.0"
description = "AI-driven research template powered by Astral tools"
readme = "README.md"
authors = [{ name = "kaito" }]
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.0.0",
    "pandas>=2.0.0",
    "matplotlib>=3.9.0",
    "pyyaml>=6.0.0",
    "pydantic>=2.0.0",
    "tqdm>=4.66.0",
    "pymupdf4llm>=0.2.9",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "ipykernel>=6.29.0",
    "ruff>=0.9.0",
    "marimo>=0.10.0",
    "poethepoet>=0.32.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ai_research_template"]

[tool.uv]
managed = true
package = true

[tool.ruff]
line-length = 88
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "UP",   # pyupgrade
    "N",    # pep8-naming
    "PL",   # pylint
    "RUF",  # ruff-specific rules
]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.knot]
# ty (formerly Red Knot) configuration
# Note: As of early 2026, configuration keys may evolve.
level = "strict"
python-version = "3.13"

[tool.poe.tasks]
exp = "python scripts/run_experiment.py"
edit = "marimo edit notebooks/analysis_sample.py"
app = "marimo run notebooks/analysis_sample.py --headless"
paper = "python scripts/convert_papers.py"
pdf2img = "python scripts/pdf_to_images.py"
check = "ruff check --fix src scripts"
format = "ruff format src scripts"
lint = ["format", "check"]
test = "pytest"
tex-sotsuron = "latexmk -cd -output-directory=out tex/sotsuron/main.tex"
tex-zemi = "latexmk -cd -output-directory=out tex/zemi/main.tex"
tex-handout = "latexmk -cd -output-directory=out tex/handout/main.tex"
tex-clean = "latexmk -C -output-directory=out tex/sotsuron/main.tex tex/zemi/main.tex tex/handout/main.tex"
